+++ final/src/main/java/shoppingcart/model/PaymentStrategy.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface PaymentStrategy {
 
-	public float makePayment(List<Item> items);
+	public double makePayment(List<Item> items);
 }
